requirements.txt
asgiref==3.2.3
Django==3.0.1
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
pytz==2019.3
sqlparse==0.3.0
mkdocs==1.0.4
